requirement.txt
praw
python-dotenv
numpy
pandas
torch
matplotlib
datasets
scikit-learn
transformers
tqdm
